Fixed selection not saving.
