+++ app/src/main/java/com/guillermonegrete/tts/importtext/visualize/VisualizeTextActivity.kt
@@ -37,6 +37,7 @@ import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.guillermonegrete.tts.EventObserver
 import com.guillermonegrete.tts.R
 import com.guillermonegrete.tts.common.compose.ExternalLinkList
+import com.guillermonegrete.tts.common.compose.ExternalLinksDialog
 import com.guillermonegrete.tts.common.models.EditNote
 import com.guillermonegrete.tts.common.models.NoteItem
 import com.guillermonegrete.tts.common.models.Span
@@ -46,7 +47,6 @@ import com.guillermonegrete.tts.db.ExternalLink
 import com.guillermonegrete.tts.importtext.epub.NavPoint
 import com.guillermonegrete.tts.importtext.visualize.model.BookChapter
 import com.guillermonegrete.tts.importtext.visualize.model.SplitPageSpan
-import com.guillermonegrete.tts.textprocessing.ExternalLinksDialog
 import com.guillermonegrete.tts.textprocessing.TextInfoDialog
 import com.guillermonegrete.tts.ui.BrightnessTheme
 import com.guillermonegrete.tts.ui.theme.AppTheme
@@ -82,8 +82,8 @@ class VisualizeTextActivity: AppCompatActivity() {
     private val addNoteDialogVisible = mutableStateOf(false)
     private val noteSheetVisible = mutableStateOf(false)
     private var linksDialogShown = mutableStateOf(false)
-    private val selectedLink = mutableIntStateOf(0)
     private val wordLinks = mutableStateOf(ExternalLinkList(emptyList()))
+    private var selectedLink = 0
 
     private var noteInfo = mutableStateOf<EditNote?>(null)
 
@@ -381,7 +381,7 @@ class VisualizeTextActivity: AppCompatActivity() {
             linksForWord.observe(this@VisualizeTextActivity) { links ->
                 wordLinks.value = ExternalLinkList(links.map(ExternalLink::toUI))
                 // If out of index, default to the first item
-                if(selectedLink.intValue >= links.size) selectedLink.intValue = 0
+                if(selectedLink >= links.size) selectedLink = 0
                 linksDialogShown.value = true
             }
 
@@ -848,7 +848,8 @@ class VisualizeTextActivity: AppCompatActivity() {
         ExternalLinksDialog(
             isShown = linksDialogShown.value,
             links = wordLinks.value,
-            selection = selectedLink.intValue,
+            selection = selectedLink,
+            onItemClick = { selectedLink = it },
             onDismiss = { linksDialogShown.value = false },
         )
