+++ app/src/main/java/com/guillermonegrete/tts/textprocessing/TextInfoDialog.kt
@@ -24,6 +24,7 @@ import androidx.viewpager2.widget.ViewPager2
 import com.google.android.material.tabs.TabLayoutMediator
 import com.guillermonegrete.tts.R
 import com.guillermonegrete.tts.common.compose.ExternalLinkList
+import com.guillermonegrete.tts.common.compose.ExternalLinksDialog
 import com.guillermonegrete.tts.common.compose.StringList
 import com.guillermonegrete.tts.common.compose.YesNoDialog
 import com.guillermonegrete.tts.common.models.Span
@@ -88,7 +89,7 @@ class TextInfoDialog: DialogFragment(), ProcessTextContract.View {
     private val wordState = mutableStateOf(WordState())
 
     private val wordLinks = mutableStateOf(ExternalLinkList(emptyList()))
-    private val selectedLink = mutableIntStateOf(0)
+    private var selectedLink = 0
 
     private val editDialogShown = mutableStateOf(false)
     private val deleteDialogShown = mutableStateOf(false)
@@ -175,7 +176,7 @@ class TextInfoDialog: DialogFragment(), ProcessTextContract.View {
 
                         EditDeleteWordDialogs(wordState)
 
-                        ExternalLinksDialog(selectedLink)
+                        LocalExternalLinksDialog()
                     }
                 }
             }
@@ -264,7 +265,7 @@ class TextInfoDialog: DialogFragment(), ProcessTextContract.View {
         presenterImp.wordLinks.observe(this) { links ->
             wordLinks.value = ExternalLinkList(links.map(ExternalLink::toUI))
             // If out of index, default to the first item
-            if(selectedLink.intValue >= links.size) selectedLink.intValue = 0
+            if(selectedLink >= links.size) selectedLink = 0
             linksDialogShown.value = true
         }
 
@@ -899,12 +900,13 @@ class TextInfoDialog: DialogFragment(), ProcessTextContract.View {
     }
 
     @Composable
-    fun ExternalLinksDialog(selectedLink: MutableIntState) {
+    fun LocalExternalLinksDialog() {
 
         ExternalLinksDialog(
             isShown = linksDialogShown.value,
             links = wordLinks.value,
-            selection = selectedLink.intValue,
+            selection = selectedLink,
+            onItemClick = { selectedLink = it},
             onDismiss = { linksDialogShown.value = false },
         )
     }
